+++ src/datasources/results/components/ResultsQueryEditor.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback } from 'react';
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
 import { ResultsDataSource } from '../ResultsDataSource';
-import { QueryType, ResultsQuery } from '../types/types';
+import { QueryType, ResultsDataSourceOptions, ResultsQuery } from '../types/types';
 import { QueryResultsEditor } from './editors/query-results/QueryResultsEditor';
 import { QueryResults } from '../types/QueryResults.types';
 import { defaultResultsQuery, defaultStepsQuery } from '../defaultQueries';
@@ -9,7 +9,7 @@ import { InlineField, RadioButtonGroup, VerticalGroup } from '@grafana/ui';
 import { QueryStepsEditor } from './editors/query-steps/QueryStepsEditor';
 import { QuerySteps } from '../types/QuerySteps.types';
 
-type Props = QueryEditorProps<ResultsDataSource, ResultsQuery>;
+type Props = QueryEditorProps<ResultsDataSource, ResultsQuery, ResultsDataSourceOptions>;
 
 export function ResultsQueryEditor({ query, onChange, onRunQuery, datasource }: Props) {
   query = datasource.prepareQuery(query);
@@ -53,6 +53,7 @@ export function ResultsQueryEditor({ query, onChange, onRunQuery, datasource }: 
         <QueryResultsEditor
           query={query as QueryResults} 
           handleQueryChange={handleQueryChange}
+          datasource={datasource.queryResultsDataSource}
         />
       )}
       {query.queryType === QueryType.Steps && (